rust_lib_still_alive

Module frb_generated

Re-exports

pub use io::*;

Modules

io 🔒

Structs

FLUTTER_RUST_BRIDGE_HANDLER

IdGenerator 🔒

MoiArc

MoiArcPoolInner

MoiArcPoolValue 🔒

StreamSink

Constants

FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH 🔒

FLUTTER_RUST_BRIDGE_CODEGEN_VERSION 🔒

Traits

CstDecode

MoiArcValue

SseDecode

SseEncode

Functions

decode_rust_opaque_moi 🔒

pde_ffi_dispatcher_primary_impl 🔒

pde_ffi_dispatcher_sync_impl 🔒

transform_result_sse 🔒

wire__crate__api__backup__export_backup_impl 🔒

wire__crate__api__backup__import_backup_impl 🔒

wire__crate__api__data__db__close_database_impl 🔒

wire__crate__api__data__db__database_path_str_impl 🔒

wire__crate__api__data__db__execute_batch_sql_impl 🔒

wire__crate__api__data__db__execute_sql_impl 🔒

wire__crate__api__data__db__export_database_impl 🔒

wire__crate__api__data__db__get_database_name_impl 🔒

wire__crate__api__data__db__import_database_impl 🔒

wire__crate__api__data__db__init_database_impl 🔒

wire__crate__api__data__db__open_database_impl 🔒

wire__crate__api__data__db__purge_database_impl 🔒

wire__crate__api__data__db__select_impl 🔒

wire__crate__api__data__db__select_one_impl 🔒

wire__crate__api__data__db__update_message_impl 🔒

wire__crate__api__data__logging__create_rust_log_stream_impl 🔒

wire__crate__api__data__logging__dispose_rust_log_stream_impl 🔒

wire__crate__api__data__logging__init_rust_logging_impl 🔒

wire__crate__api__integrations__public_traits__connect_integration_account_impl 🔒

wire__crate__api__integrations__public_traits__delete_integration_account_impl 🔒

wire__crate__api__integrations__public_traits__discover_integration_destinations_impl 🔒

wire__crate__api__integrations__public_traits__load_all_integrations_impl 🔒

wire__crate__api__integrations__public_traits__send_integration_message_impl 🔒

wire__crate__api__integrations__public_traits__test_integration_account_impl 🔒

wire__crate__api__integrations__public_traits__test_integration_destination_impl 🔒

wire__crate__api__main__greet_impl 🔒

wire__crate__api__main__init_app_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_auto_accessor_get_config_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_auto_accessor_get_key_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_auto_accessor_get_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_auto_accessor_set_config_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_auto_accessor_set_key_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_auto_accessor_set_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_cancel_timer_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_delete_timer_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_pause_timer_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_replace_active_timer_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_resume_timer_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_start_timer_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_start_warning_run_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_trigger_manual_expire_impl 🔒

wire__crate__api__timer__active_timer__ActiveTimer_update_run_state_impl 🔒

wire__crate__api__timer__active_timer__get_unique_timer_id_impl 🔒

wire__crate__api__timer__active_timer__persist_timer_impl 🔒

wire__crate__api__timer__active_timer__reconcile_active_timer_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_audio_recording_enabled_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_contacts_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_created_at_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_custom_email_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_custom_sms_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_duration_secs_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_grace_period_secs_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_integrations_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_location_collection_interval_secs_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_location_sharing_enabled_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_message_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_name_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_password_hash_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_password_protected_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_route_sharing_enabled_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_get_updated_at_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_audio_recording_enabled_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_contacts_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_created_at_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_custom_email_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_custom_sms_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_duration_secs_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_grace_period_secs_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_integrations_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_location_collection_interval_secs_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_location_sharing_enabled_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_message_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_name_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_password_hash_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_password_protected_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_route_sharing_enabled_impl 🔒

wire__crate__api__timer__config__TimerConfig_auto_accessor_set_updated_at_impl 🔒

wire__crate__api__timer__config__TimerConfig_new_impl 🔒

wire__crate__api__timer__run__timer_run_evaluate_state_impl 🔒

wire__crate__api__timer__run__timer_run_is_active_impl 🔒

wire__crate__api__timer__state__timer_state_as_str_impl 🔒

wire__crate__api__timer__state__timer_state_state_from_str_impl 🔒

Type Aliases

MoiArcPool

MoiObjectId 🔒

RustAutoOpaque

Usually this is unneeded, and just write down arbitrary types. However, when you need arbitrary types at places that are not supported yet, use RustOpaqueOpaque<YourArbitraryType>.

RustAutoOpaqueMoi

Please refer to RustAutoOpaque for doc.

RustOpaque

A wrapper to support arbitrary Rust types.

RustOpaqueMoi

Please refer to RustOpaque for doc.